learning creating apps using app factory
