+++ _wsgi.py
@@ -1,0 +1,7 @@
+from blog.app import app
+
+if __name__ == '__main__':
+    app.run(
+        host='0.0.0.0',
+        debug=True
+    )
