+++ wsgi.py
@@ -1,7 +1,6 @@
-from blog.app import app
+"""
+Здесь второй вариант запуска приложени через фабрику приложений
+"""
+from blog.app import create_app
 
-if __name__ == '__main__':
-    app.run(
-        host='0.0.0.0',
-        debug=True
-    )
+app = create_app()
